myapp/requirements.txt
Django==2.1.8
djangorestframework==3.11.0
psycopg2-binary==2.8.5
pytz==2019.3
sqlparse==0.3.1
djangorestframework-simplejwt==4.6.0
django-cors-headers
django-sendgrid-v5
django_rest_passwordreset==1.3.0
reportlab==3.6.12
python-dateutil==2.8.2